refactor: add NonProfitsList component
